added confirmation box and compare booked date

+++ staff/admin/detailAsset.php
@@ -37,7 +37,7 @@ $result = query($query);
                     <td><?= $res['asset_brand'] ?></td>
                     <td>
                         <a href="">Edit</a> |  <!-- TO DO: buat edit page -->
-                        <a href="../backend/hapusAssetFunction.php?id=<?= htmlspecialchars($res['asset_id']) ?>">Hapus</a>   <!-- DONE: kalau dihapus asetnya gmn? -->
+                        <a href="../backend/hapusAssetFunction.php?id=<?= htmlspecialchars($res['asset_id']) ?>" onclick="return confirm('Asset akan dihapus?');">Hapus</a>   <!-- DONE: kalau dihapus asetnya gmn? -->
                     </td>
                     <td>
                         <?php
